+++ src/features/persoon/persoon-details/use-persoon-by-bsn.ts
@@ -1,18 +1,10 @@
 import { enforceOneOrZero, ServiceResult } from "@/services";
 import { searchPersonen } from "@/services/brp";
-import type { Ref } from "vue";
 
 export function usePersoonByBsn(
-  getBsn: () => string | undefined,
-  useOk2: Ref<boolean | null>,
+  getBsn: () => string | undefined
 ) {
-  // const getCacheKey = () => {
-  //   if (useOk2.value === null) {
-  //     return "";
-  //   }
-  //   const bsn = getBsn();
-  //   return bsn ? "persoon" + bsn : "";
-  // };
+
   const fetcher = (bsn: string) => {
     return searchPersonen({ bsn }).then(enforceOneOrZero);
   };
